Simplify board styles, AI composer and shared task rules
