+++ shared/api-fields.mjs
@@ -1,0 +1,142 @@
+import { TASK_STATUSES, isTaskStatus, isTaskPriority } from "./domain.mjs";
+
+export class ApiError extends Error {
+  constructor(status, code, message, details) {
+    super(message);
+    this.name = "ApiError";
+    this.status = status;
+    this.code = code;
+    this.details = details;
+  }
+}
+
+export function assertPlainObject(value) {
+  if (value === null || typeof value !== "object" || Array.isArray(value)) {
+    throw new ApiError(400, "INVALID_BODY", "Request body must be a JSON object");
+  }
+}
+
+export function assertAllowedKeys(value, allowed) {
+  const unknown = Object.keys(value).filter((key) => !allowed.has(key));
+  if (unknown.length > 0) {
+    throw new ApiError(400, "UNKNOWN_FIELD", `Unknown field${unknown.length === 1 ? "" : "s"}: ${unknown.join(", ")}`);
+  }
+}
+
+export function stringField(value, name, { required = false, nullable = false, maxLength } = {}) {
+  if (value === undefined) {
+    if (required) {
+      throw new ApiError(400, "INVALID_FIELD", `'${name}' is required`);
+    }
+    return undefined;
+  }
+  if (nullable && value === null) {
+    return null;
+  }
+  if (typeof value !== "string") {
+    throw new ApiError(400, "INVALID_FIELD", `'${name}' must be a string${nullable ? " or null" : ""}`);
+  }
+  const normalized = value.trim();
+  if (required && normalized.length === 0) {
+    throw new ApiError(400, "INVALID_FIELD", `'${name}' cannot be empty`);
+  }
+  if (normalized.length > maxLength) {
+    throw new ApiError(400, "INVALID_FIELD", `'${name}' cannot exceed ${maxLength} characters`);
+  }
+  return normalized;
+}
+
+export function parseDueDate(value, name = "dueDate") {
+  const date = stringField(value, name, { nullable: true, maxLength: 10 });
+  if (date !== null && date !== undefined && !/^\d{4}-\d{2}-\d{2}$/.test(date)) {
+    throw new ApiError(400, "INVALID_FIELD", `'${name}' must use YYYY-MM-DD`);
+  }
+  return date;
+}
+
+export function parseRecurrence(value) {
+  if (value === null) return null;
+  assertPlainObject(value);
+  assertAllowedKeys(value, new Set(["interval", "unit"]));
+  if (!Number.isSafeInteger(value.interval) || value.interval < 1 || value.interval > 365) {
+    throw new ApiError(400, "INVALID_FIELD", "'recurrence.interval' must be an integer from 1 to 365");
+  }
+  if (!["day", "week", "month", "year"].includes(value.unit)) {
+    throw new ApiError(400, "INVALID_FIELD", "'recurrence.unit' must be day, week, month, or year");
+  }
+  return { interval: value.interval, unit: value.unit };
+}
+
+export function parseSortOrder(value) {
+  if (typeof value !== "number" || !Number.isFinite(value) || Math.abs(value) > 1_000_000_000_000) {
+    throw new ApiError(400, "INVALID_FIELD", "'sortOrder' must be a finite number between -1000000000000 and 1000000000000");
+  }
+  return value;
+}
+
+export function parseLabels(value) {
+  if (!Array.isArray(value) || value.length > 20) {
+    throw new ApiError(400, "INVALID_FIELD", "'labels' must be an array with at most 20 entries");
+  }
+  const labels = value.map((label) => {
+    if (typeof label !== "string") {
+      throw new ApiError(400, "INVALID_FIELD", "Every label must be a string");
+    }
+    const normalized = label.trim();
+    if (normalized.length === 0 || normalized.length > 64) {
+      throw new ApiError(400, "INVALID_FIELD", "Labels must contain 1 to 64 characters");
+    }
+    return normalized;
+  });
+  if (new Set(labels).size !== labels.length) {
+    throw new ApiError(400, "INVALID_FIELD", "Labels must be unique");
+  }
+  return labels;
+}
+
+export function parseStatus(value, fallback) {
+  const result = value ?? fallback;
+  if (!isTaskStatus(result)) {
+    throw new ApiError(400, "INVALID_FIELD", `'status' must be one of: ${TASK_STATUSES.join(", ")}`);
+  }
+  return result;
+}
+
+export function parsePriority(value, fallback) {
+  const result = value ?? fallback;
+  if (!isTaskPriority(result)) {
+    throw new ApiError(400, "INVALID_FIELD", "'priority' must be none, urgent, high, medium, or low");
+  }
+  return result;
+}
+
+export function slugify(value) {
+  return value.toLowerCase().replace(/[^a-z0-9]+/g, "-").replace(/^-+|-+$/g, "").slice(0, 64);
+}
+
+export function parseProjectLabel(body) {
+  assertPlainObject(body);
+  assertAllowedKeys(body, new Set(["label"]));
+  return stringField(body.label, "label", { required: true, maxLength: 64 });
+}
+
+export function parseThreadId(value) {
+  if (value === undefined) return undefined;
+  return stringField(value, "threadId", { required: true, maxLength: 256 });
+}
+
+export function parseAssigneeTarget(value) {
+  if (value === undefined) return undefined;
+  if (value !== "current-user" && value !== "codex-agent") {
+    throw new ApiError(400, "INVALID_FIELD", "'assigneeTarget' must be current-user or codex-agent");
+  }
+  return value;
+}
+
+export function parseRelationOrigin(value) {
+  if (value === undefined) return undefined;
+  if (value !== "manual" && value !== "mention") {
+    throw new ApiError(400, "INVALID_FIELD", "'origin' must be manual or mention");
+  }
+  return value;
+}
