--- test/board-interactions.test.mjs
@@ -1,181 +1,0 @@
-import assert from "node:assert/strict";
-import { readFile } from "node:fs/promises";
-import { test } from "node:test";
-
-const appSource = await readFile(new URL("../web/src/App.tsx", import.meta.url), "utf8");
-const boardColumnSource = await readFile(new URL("../web/src/components/BoardColumn.tsx", import.meta.url), "utf8");
-const apiSource = await readFile(new URL("../web/src/api.ts", import.meta.url), "utf8");
-const styles = await readFile(new URL("../web/src/styles.css", import.meta.url), "utf8");
-const detailSource = await readFile(new URL("../web/src/components/TaskDetail.tsx", import.meta.url), "utf8");
-const editorSource = await readFile(new URL("../web/src/components/TaskEditor.tsx", import.meta.url), "utf8");
-const labelPickerSource = await readFile(new URL("../web/src/components/LabelPicker.tsx", import.meta.url), "utf8");
-const contextMenuSource = await readFile(new URL("../web/src/components/TaskContextMenu.tsx", import.meta.url), "utf8");
-const cardSource = await readFile(new URL("../web/src/components/TaskCard.tsx", import.meta.url), "utf8");
-const filterSource = await readFile(new URL("../web/src/taskFilters.ts", import.meta.url), "utf8");
-const typesSource = await readFile(new URL("../web/src/types.ts", import.meta.url), "utf8");
-const composerSource = await readFile(new URL("../web/src/components/InlineMediaComposer.tsx", import.meta.url), "utf8");
-
-function taskStatuses() {
-  const match = typesSource.match(/export const TASK_STATUSES = (\[[\s\S]*?\]) as const/);
-  assert.ok(match);
-  return [...match[1].matchAll(/"([^"]+)"/g)].map((entry) => entry[1]);
-}
-
-test("dragging previews the insertion rank before committing it", () => {
-  assert.match(boardColumnSource, /function findDropBefore/);
-  assert.match(boardColumnSource, /clientY < card\.getBoundingClientRect\(\)\.top \+ card\.offsetHeight \/ 2/);
-  assert.match(boardColumnSource, /onDrop\(status, taskId, findDropBefore/);
-  assert.match(boardColumnSource, /function getTaskDragShift/);
-  assert.match(boardColumnSource, /shift -= dragDistance/);
-  assert.match(boardColumnSource, /shift \+= dragDistance/);
-  assert.match(boardColumnSource, /dragShift=\{dragShift\}/);
-  assert.match(styles, /\.task-card\.is-dragging \{[\s\S]*?opacity: 0/);
-  assert.doesNotMatch(styles, /\.task-card\.is-dragging \{[^}]*pointer-events: none/);
-  assert.match(styles, /transform 160ms cubic-bezier/);
-  assert.match(appSource, /beforeTaskId: string \| null = null/);
-  assert.match(appSource, /\(previousTask\.sortOrder \+ nextTask\.sortOrder\) \/ 2/);
-  assert.match(appSource, /currentOrder\.every\(\(candidate, index\) => candidate\.id === desiredOrder\[index\]\.id\)/);
-  assert.match(appSource, /setTasks\(\(current\) => sortTasks\(current\.map/);
-  assert.match(appSource, /setSettlingTaskId\(task\.id\)/);
-  assert.match(styles, /\.task-card\.is-settling \{[\s\S]*?task-card-settle 200ms/);
-});
-
-test("text selection is reserved for editable fields", () => {
-  assert.match(styles, /body \{[^}]*user-select: none/);
-  assert.match(styles, /input,[\s\S]*?textarea,[\s\S]*?\[contenteditable="true"\][\s\S]*?user-select: text/);
-});
-
-test("main issue cards stay compact while sidebar cards show ownership and creation time", () => {
-  assert.match(cardSource, /projectName[\s\S]*?className="project-chip"/);
-  assert.match(cardSource, /variant === "sidebar" && \([\s\S]*?className="sidebar-card-creator"/);
-  assert.match(cardSource, /<AssigneeControl[\s\S]*?<span>\{createdDate\(task\.createdAt, locale, text\)\}<\/span>/);
-  assert.doesNotMatch(styles, /\.card-footer|\.created-at/);
-  assert.match(styles, /\.project-chip/);
-  assert.match(styles, /\.task-card \{[\s\S]*?min-height: 80px;[\s\S]*?gap: 6px;[\s\S]*?padding: 7px 8px/);
-  assert.match(detailSource, /currentTask\.createdAt/);
-});
-
-test("native select options remain readable in dark theme", () => {
-  assert.match(styles, /:root\[data-theme="dark"\] select \{[\s\S]*?color-scheme: dark/);
-  assert.match(styles, /:root\[data-theme="dark"\] select option \{[\s\S]*?background-color: var\(--surface-raised\);[\s\S]*?color: var\(--text-primary\)/);
-  assert.match(styles, /:root\[data-theme="dark"\] select option:checked \{[\s\S]*?background-color: var\(--surface-active\)/);
-});
-
-test("each status column remains a drop target for the full board height", () => {
-  assert.match(styles, /\.board \{[\s\S]*?align-items: stretch;[\s\S]*?height: 100%/);
-  assert.match(styles, /\.board-column \{[\s\S]*?display: flex;[\s\S]*?flex-direction: column;[\s\S]*?height: 100%/);
-  assert.match(styles, /\.column-list \{[\s\S]*?flex: 1 0 auto;[\s\S]*?min-height: calc\(100% - 48px\)/);
-});
-
-test("the issue board has no shared vertical scroll and each status column scrolls below a sticky heading", () => {
-  assert.match(styles, /\.board-scroll \{[\s\S]*?overflow-x: auto;[\s\S]*?overflow-y: hidden;[\s\S]*?overscroll-behavior-y: none/);
-  assert.match(styles, /\.board-column \{[\s\S]*?overflow-x: hidden;[\s\S]*?overflow-y: auto;[\s\S]*?overscroll-behavior-y: contain/);
-  assert.match(styles, /\.column-header \{[\s\S]*?position: sticky;[\s\S]*?top: 0;[\s\S]*?background: var\(--board-column-surface\)/);
-});
-
-test("the complete issue status set shares one ordered source", () => {
-  assert.deepEqual(taskStatuses(), [
-    "backlog",
-    "todo",
-    "in_progress",
-    "in_review",
-    "blocked",
-    "done",
-    "canceled",
-  ]);
-  assert.match(boardColumnSource, /backlog: \{ label: "待立项", tone: "backlog" \}/);
-  assert.match(boardColumnSource, /todo: \{ label: "等待认领", tone: "todo" \}/);
-  assert.match(boardColumnSource, /in_progress: \{ label: "处理中", tone: "progress" \}/);
-  assert.match(boardColumnSource, /in_review: \{ label: "等你确认", tone: "review" \}/);
-  assert.match(boardColumnSource, /blocked: \{ label: "遇到阻碍", tone: "blocked" \}/);
-  assert.match(boardColumnSource, /done: \{ label: "完成", tone: "done" \}/);
-  assert.match(boardColumnSource, /canceled: \{ label: "取消", tone: "canceled" \}/);
-  assert.doesNotMatch(cardSource, /STATUS_ORDER/);
-  assert.match(detailSource, /TASK_STATUSES\.map\(\(status\) =>/);
-  assert.match(editorSource, /TASK_STATUSES\.map\(\(value\) =>/);
-  assert.match(contextMenuSource, /TASK_STATUSES\.map\(\(status, index\) =>/);
-});
-
-test("review, blocked and canceled statuses round-trip through filter URLs", () => {
-  const statuses = taskStatuses();
-  const selected = ["in_review", "blocked", "canceled"];
-  const url = new URL("http://taskboard.local/");
-  url.searchParams.set("status", selected.join(","));
-  const restored = url.searchParams.get("status").split(",").filter((status) => statuses.includes(status));
-
-  assert.deepEqual(restored, selected);
-  assert.match(filterSource, /filters\.statuses\.join\(","\)/);
-  assert.match(filterSource, /\.split\(","\)\.filter\(isTaskStatus\)/);
-  assert.match(filterSource, /TASK_STATUSES\.includes\(value as TaskStatus\)/);
-});
-
-test("the column surface wraps its heading and issue list", () => {
-  assert.match(styles, /\.board-column \{[\s\S]*?--board-column-surface: var\(--column-header\)[\s\S]*?background: var\(--board-column-surface\)/);
-  assert.match(styles, /\.column-header \{[\s\S]*?background: var\(--board-column-surface\)/);
-  assert.match(styles, /\.column-list \{[\s\S]*?padding: 8px 8px 8px/);
-});
-
-test("common issue mutations enter a Linear-style undo queue", () => {
-  assert.match(appSource, /const undoStackRef = useRef<UndoOperation\[]>/);
-  assert.match(appSource, /event\.key\.toLowerCase\(\) === "z"/);
-  assert.match(appSource, /event\.metaKey \|\| event\.ctrlKey/);
-  assert.match(appSource, /function pushUndo/);
-  assert.match(appSource, /function performUndo[\s\S]*?await operation\.undo\(\)/);
-  assert.match(appSource, /void performUndo\(\)/);
-  assert.match(appSource, /moveTask\(task, destination, beforeTaskId, true\)/);
-  assert.match(appSource, /className="toast undo-toast"/);
-  assert.match(appSource, />\s*\{text\("撤回", "Undo"\)\} <kbd>\{undoShortcut\}<\/kbd>/);
-  assert.match(appSource, /restoreTaskRequest\(archived\)/);
-  assert.match(apiSource, /export async function restoreTask/);
-});
-
-test("issues expose processing conversations without manual binding", () => {
-  assert.match(detailSource, /在新对话打开/);
-  assert.match(detailSource, /onOpenInThread\(currentTask\)/);
-  assert.doesNotMatch(appSource, /detail-thread-button/);
-  assert.doesNotMatch(detailSource, /输入对话 ID|解除 Codex 对话绑定|>绑定</);
-  assert.doesNotMatch(editorSource, /对话 ID|linkedThreadId/);
-  assert.match(detailSource, /currentTask\.threadBinding \|\| currentTask\.legacyLocalThreadId/);
-  assert.doesNotMatch(detailSource, /currentTask\.threadIds/);
-  assert.match(detailSource, /<strong>\{text\("查看对话", "View conversation"\)\}<\/strong>/);
-  assert.doesNotMatch(detailSource, /className="conversation-thread-id">\{threadId\}/);
-  assert.doesNotMatch(detailSource, /shortThreadId/);
-  assert.doesNotMatch(detailSource, /detail-property-label">Codex/);
-  assert.match(detailSource, /comment\.threadBinding \|\| comment\.legacyLocalThreadId/);
-  assert.match(detailSource, /onOpenLegacyLocalThread\(comment\.legacyLocalThreadId!\)/);
-  assert.doesNotMatch(detailSource, /compact/);
-  assert.doesNotMatch(styles, /issue-conversation-link\.compact/);
-  assert.match(detailSource, /\.\.\.developmentOptions\.map\(\(context\) => \(\{/);
-  assert.match(detailSource, /context\.type === "branch"[\s\S]*?<BranchIcon[\s\S]*?<LinearIcon name="folder"/);
-  assert.match(detailSource, /developmentContext/);
-  assert.doesNotMatch(detailSource, /placeholder="绑定分支/);
-  assert.doesNotMatch(contextMenuSource, /打开关联 Codex 对话/);
-  assert.match(contextMenuSource, /onOpenInThread/);
-});
-
-test("comments upload and render their own attachments in the content flow", () => {
-  assert.match(apiSource, /export async function uploadCommentAttachment/);
-  assert.match(apiSource, /\/api\/comments\/\$\{encodeURIComponent\(commentId\)\}\/attachments/);
-  assert.match(detailSource, /commentInlineFiles/);
-  assert.match(detailSource, /uploadCommentAttachment\(comment\.id, file\.file, "attachment"\)/);
-  assert.match(detailSource, /resolveInlineAttachmentMarkdown/);
-  assert.match(detailSource, /createInlineMediaSegments\(comment\.body, referenceTasks, comment\.attachments\)/);
-  assert.match(detailSource, /attachments=\{comment\.attachments\}/);
-  assert.match(detailSource, /onOpenAttachment=\{handleAttachmentDownload\}/);
-  assert.match(composerSource, /className="inline-media-attachment"/);
-});
-
-test("issue creation and detail share one searchable, creatable label picker", () => {
-  assert.match(editorSource, /<LabelPicker/);
-  assert.match(detailSource, /<LabelPicker/);
-  assert.match(appSource, /<TaskDetail[\s\S]*?availableLabels=\{availableLabels\}/);
-  assert.match(detailSource, /selectedLabels=\{currentTask\.labels\}/);
-  assert.match(detailSource, /saveTask\(\{ labels: nextLabels \}, "labels"\)/);
-  assert.doesNotMatch(detailSource, /标签，以逗号分隔|function saveLabels|labels\.split/);
-  assert.match(labelPickerSource, /availableLabels\.filter/);
-  assert.match(labelPickerSource, /selectedLabels\.includes\(label\)/);
-  assert.match(labelPickerSource, /text\(`创建 “\$\{normalizedSearch\}”`, `Create “\$\{normalizedSearch\}”`\)/);
-  assert.match(labelPickerSource, /labelPresentation\(normalizedSearch, language\)\.color/);
-  assert.match(labelPickerSource, /aria-multiselectable="true"/);
-  assert.match(styles, /\.detail-label-picker \.label-popover/);
-});
